Move Proxy to own section and display DNS warnings
