+++ src/popup/views/Home.vue
@@ -2,8 +2,8 @@
 import { provide } from 'vue';
 
 import RecommendationsCarousel from '@/components/RecommendationsCarousel.vue';
-import WarningsCarousel from '@/components/WarningsCarousel.vue';
 import ConnectionDetails from '@/components/ConnectionDetails/ConnectionDetails.vue';
+import ProxyDetails from '@/components/ProxyDetails/ProxyDetails.vue';
 
 import useConnection, { ConnectionKey } from '@/composables/useConnection';
 
@@ -12,7 +12,7 @@ provide(ConnectionKey, { connection, isLoading, isError });
 </script>
 
 <template>
-  <WarningsCarousel />
   <ConnectionDetails />
+  <ProxyDetails />
   <RecommendationsCarousel />
 </template>
